+++ Assets/Prefabs/EnemyPrefabs/FlyingMeleeEnemy.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 6f1f6815f883b224799cbf58c935c9a2
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
