+++ Assets/Prefabs/WeaponPrefabs/Flamethrower/Flame.prefab
@@ -170,3 +170,7 @@ MonoBehaviour:
   enemyLayer:
     serializedVersion: 2
     m_Bits: 256
+  flyingEnemyLayer:
+    serializedVersion: 2
+    m_Bits: 512
+  pushBackForce: 100
