--- Assets/Prefabs/WeaponPrefabs/Flamethrower/SlowEffect.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 3834a29ad4b510641884470519aa131d
