+++ Assets/Scripts/Weapons/FlameThrower/SlowEffect.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 3834a29ad4b510641884470519aa131d
