+++ ProjectSettings/TagManager.asset
@@ -15,7 +15,7 @@ TagManager:
   - Player
   - Projectile
   - Enemy
-  - 
+  - FlyingEnemy
   - 
   - 
   -
